common/Finisher: use `std::lock_guard` instead of `std::unique_lock`
`std::lock_guard` is all we need here, and the added complexity of
`std::unique_lock` is not used and is usually optimized away by the
compiler.  Using `std::lock_guard` directly will reduce the amount of
work that the optimizer needs to do and saves some build CPU cycles.

Signed-off-by: Max Kellermann <max.kellermann@ionos.com>