osd: Torn write protection for Direct Reads
It is possible for direct reads to query two seperate shards and get different versions of the object for each shard when using direct reads.

To solve this we add a get_internal_version op to tell us the version of the object on that shard and submit that in the same transaction as the read so we can ensure the versions are what we expect. If we have a mismatch, we resubmit the read through the primary path.

Signed-off-by: Jon Bailey <jonathan.bailey1@ibm.com>
Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>